Refactor: simplify path resolution
