+++ winbooks-api-extra/src/main/java/be/valuya/winbooks/api/accountingtroll/cache/AccountingManagerCache.java
@@ -1,17 +1,14 @@
 package be.valuya.winbooks.api.accountingtroll.cache;
 
-import be.valuya.accountingtroll.AccountingEventListener;
 import be.valuya.accountingtroll.domain.ATAccount;
 import be.valuya.accountingtroll.domain.ATAccountingEntry;
 import be.valuya.accountingtroll.domain.ATBookPeriod;
 import be.valuya.accountingtroll.domain.ATBookYear;
 import be.valuya.accountingtroll.domain.ATDocument;
 import be.valuya.accountingtroll.domain.ATThirdParty;
-import be.valuya.accountingtroll.event.AccountingEventHandler;
 import be.valuya.jbooks.model.WbAccount;
 import be.valuya.jbooks.model.WbBookYearFull;
 import be.valuya.jbooks.model.WbClientSupplier;
-import be.valuya.jbooks.model.WbDbkType;
 import be.valuya.jbooks.model.WbEntry;
 import be.valuya.jbooks.model.WbPeriod;
 import be.valuya.winbooks.api.accountingtroll.converter.ATAccountConverter;
@@ -140,7 +137,7 @@ public class AccountingManagerCache {
             return;
         }
 
-        int accountNumberLength = extraService.getAccountNumberLength(fileConfiguration);
+        int accountNumberLength = extraService.getAccountNumberLengthFromParamsTable(fileConfiguration);
         accountsByCode = extraService.streamAcf(fileConfiguration)
                 .filter(this::isValidAccount)
                 .map(wbAccount -> atAccountConverter.convertToTrollAcccount(wbAccount, accountNumberLength))
@@ -212,8 +209,7 @@ public class AccountingManagerCache {
             cacheDocuments();
         }
 
-        AccountingEventListener accountingEventListener = new AccountingEventHandler();
-        accountingEntries = extraService.streamAct(fileConfiguration, accountingEventListener)
+        accountingEntries = extraService.streamAct(fileConfiguration)
                 .filter(this::isValidAccountingEntry)
                 .map(atAccountingEntryConverter::convertToTrollAccountingEntry)
                 .map(e -> this.linkEntryDocument(e, documentMatchingMode))
@@ -244,9 +240,8 @@ public class AccountingManagerCache {
         if (documentsByCacheKey != null) {
             return;
         }
-        AccountingEventListener accountingEventListener = new AccountingEventHandler();
         documentsByCacheKey = streamWbBookYearFulls()
-                .flatMap(bookYear -> extraService.streamBookYearDocuments(fileConfiguration, bookYear, accountingEventListener))
+                .flatMap(bookYear -> extraService.streamBookYearDocuments(fileConfiguration, bookYear))
                 .map(atDocumentConverter::convertDocument)
                 .collect(Collectors.toMap(
                         ATDocumentCacheKey::new,
