+++ winbooks-api-extra/src/main/java/be/valuya/winbooks/api/extra/event/WinbooksEvent.java
@@ -1,4 +1,4 @@
-package be.valuya.winbooks.api.extra;
+package be.valuya.winbooks.api.extra.event;
 
 import java.util.Arrays;
 import java.util.List;
