+++ winbooks-api-extra/src/main/java/be/valuya/winbooks/api/extra/event/WinbooksEventType.java
@@ -1,4 +1,4 @@
-package be.valuya.winbooks.api.extra;
+package be.valuya.winbooks.api.extra.event;
 
 public enum WinbooksEventType {
