+++ winbooks-api-extra/src/main/java/be/valuya/winbooks/api/extra/reader/DbfException.java
@@ -1,4 +1,4 @@
-package be.valuya.winbooks.api.extra;
+package be.valuya.winbooks.api.extra.reader;
 
 /**
  *
