+++ winbooks-api-extra/src/main/java/be/valuya/winbooks/api/extra/reader/WbAccountDbfReader.java
@@ -1,4 +1,4 @@
-package be.valuya.winbooks.api.extra;
+package be.valuya.winbooks.api.extra.reader;
 
 import be.valuya.jbooks.model.WbAccount;
 import net.iryndin.jdbf.core.DbfField;
